+++ CHANGELOG.md
@@ -5,5 +5,5 @@
 - Bump minimal requirements to make the plugin compatible with newer IDE versions.
 
 ### Fixed
-- Plugin not working with Intellij IDEA 2024.3
+- Plugin not working with Intellij IDEA 2024.3.1
 - Removed unused import